% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{damage}
\alias{damage}
\title{Generate damage art}
\usage{
damage(seed, filename = NULL, pixels = 3000)
}
\arguments{
\item{seed}{integer seed specifying piece number}

\item{filename}{file to save the output to}

\item{pixels}{resolution}
}
\value{
art
}
\description{
Generate damage art
}
